rajkaj/requirements.txt
pyautogui
easyocr
opencv-python
selenium
python-dotenv
pycopy-webbrowser